Cast away a warning
git-svn-id: https://svn.blender.org/svnroot/verse/trunk@4401 49c17534-1d02-dc11-b283-0030488c597c

+++ e_storage_geometry.c
@@ -520,7 +520,7 @@ void callback_send_g_layer_create(void *user_data, VNodeID node_id, VLayerID lay
 	if(layer_id >= node->layer_allocated)
 	{
 		node->layers = realloc(node->layers, (sizeof *node->layers) * (layer_id + 8));
-		for(;node->layer_allocated < layer_id + 8; node->layer_allocated++)
+		for(;node->layer_allocated < (uint) layer_id + 8; node->layer_allocated++)
 		{
 			node->layers[node->layer_allocated].name[0] = 0;
 			node->layers[node->layer_allocated].type = -1;
